move constants.js even further away from the src/ dir to reinforce it's effectively being templated by cloudflare/esbuild
