+++ packages/api/package.json
@@ -10,7 +10,8 @@
     "dev": "miniflare --watch --debug",
     "clean": "./scripts/cli.js db --clean && rm -rf docker/compose",
     "test": "tsc && playwright-test \"test/**/*.spec.js\" --sw src/index.js",
-    "db-types": "./scripts/cli.js db-types"
+    "db-types": "./scripts/cli.js db-types",
+    "start": "node --experimental-json-modules ./src/index.js"
   },
   "author": "Hugo Dias <hugomrdias@gmail.com> (hugodias.me)",
   "license": "(Apache-2.0 OR MIT)",
@@ -20,7 +21,7 @@
     "@ipld/car": "^3.1.20",
     "@ipld/dag-cbor": "^6.0.13",
     "@ipld/dag-pb": "^2.1.16",
-    "@magic-sdk/admin": "^1.3.0",
+    "@magic-sdk/admin": "^1.4.1",
     "@nftstorage/ipfs-cluster": "^5.0.1",
     "@supabase/postgrest-js": "^0.34.1",
     "ipfs-car": "^0.6.1",
