+++ packages/api/src/routes/nfts-check.js
@@ -1,6 +1,6 @@
 import { JSONResponse } from '../utils/json-response.js'
 import { HTTPError } from '../errors.js'
-import { secrets, database } from '../constants/constants.js'
+import { secrets, database } from '../../constants/constants.js'
 import { DBClient } from '../utils/db-client.js'
 import { parseCid } from '../utils/utils.js'
 import { toCheckNftResponse } from '../utils/db-transforms.js'
